requirements.txt
torch
numpy
pandas
pyyaml
rdkit-pypi
